agnoWorkflows/pyproject.toml
[project]
name = "agnoworkflows"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agno>=2.3.13",
    "ddgs>=9.9.3",
    "fastapi>=0.124.4",
    "googlesearch-python>=1.3.0",
    "openai>=2.11.0",
    "python-dotenv>=1.2.1",
]
